role-aware-task-translator/backend/requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
#pinecone-client
openai
requests
